UnreadableBytes => long

+++ Jint.DebugAdapter/Protocol/Responses/ReadMemoryResponse.cs
@@ -3,7 +3,7 @@
     public class ReadMemoryResponse : ProtocolResponseBody
     {
         public string Address { get; set; }
-        public int? UnreadableBytes { get; set; }
+        public long? UnreadableBytes { get; set; }
         public string Data { get; set; }
     }
 }
